create a working demo!
Drag the sliders on the "Filtered" window to adjust the ranges for acceptable
hue, sat, and val levels.
Drag the "Blur" slider on the main window to control the strength of the
Gaussian blur applied to the input feed. The blur helps to remove irregularities
in the image. One "unit" on the slider equals 2px on the Gaussian blur, starting
at 1px (basically blur size=(slider*2)+1).

TODO: add some limits on the slider to prevent you from creating impossible
situations (e.g. max lower than min) and clean up some of the code that doesn't
need to be there.