Extract some functionality into header file
Thresholding and color conversion operations have been moved to utils.hpp, as
well as functions to set up and monitor trackbars for easy use in finding values
to use for the threshold operation.

This makes the code a lot simpler and makes it easy to implement the same
functionality in the StillImageTesting.cpp program without rewriting a bunch of
the same code.