Create script to generate TFRecord files
TFRecord files are required by tensorflow to format the data needed to train the
neural network. This script will generate them using the images and the CSV.

Additionally, a blank line added to the CSV by a bug in the dataset tool
application was removed. This blank line breaks the generate_tfrecord.py script.