Finalize training configuration
Training actually works now!
NOTE: The model_main.py in scripts/ is a modified version of the one provided by
the object detection API models. It mainly just contains changes to the
parameters that control number of steps to log and save after. It is
automatically copied into the correct place by the install_models.sh script.

Also, a slightly large change to make training work was to switch to Python 2
instead of 3 because the Object Detection API doesn't like 3 apparently.