Convert project back to C++
Yeah... I know.
One of the teams we need to have this program be interoperable with is using the
C++ bindings of OpenCV, so we've converted the program back to C++. The weird
error seems to be gone. Not quite sure what was going on there.

I found a lot of this C++ program from the internet and a lot of cleanup needs
to be done on it still. More to come in further commits.